--- web/utils/types.ts
@@ -1,9 +1,0 @@
-export type Poem = {
-    title: string,
-    subtitle:string,
-    date: string,
-    content: string,
-    archive:"true" | "false",
-    slug:string ,
-    filename?:string,
-  }
